% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{data_interactive}
\alias{data_interactive}
\title{data_interactive Function}
\usage{
data_interactive()
}
\arguments{
\item{data_interactive}{A function}
}
\description{
This function sets the required D3 variables for the interactive network view
}
\keyword{D3}
\keyword{interactive}
\keyword{network}
\keyword{networkD3}
