% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{data_load}
\alias{data_load}
\title{Load data}
\usage{
data_load()
}
\arguments{
\item{data_load}{A function}
}
\description{
This function reads data from the csv files for nodes and links
}
\examples{

}
\keyword{csv}
\keyword{data}
