% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Draw Function}
\usage{
draw()
}
\value{
the graph
}
\description{
This function lets user decide what kind of visual representation \cr
they want to use, for now it is limited to \cr
plot, label plot, interactive and Sankey network and shiny representation.
}
