% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\docType{data}
\name{edgesCSV}
\alias{edgesCSV}
\title{Load data}
\format{An object of class \code{character} of length 1.}
\usage{
data_load()
}
\value{
None
}
\description{
This function asks user to specify the location of the data for edges and
for nodes separately in their local computer and reads the specified CSV
formatted data into R if user doesn't specify a location for the data,
example data will be loaded.
Please answer the questions with their full location like:
"/Users/JaneDoe/Desktop/BCB410/Rview/inst/extdata/data_edges.csv"
}
\keyword{datasets}
