% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSet.R
\docType{data}
\name{igraphLib}
\alias{igraphLib}
\title{plotSet}
\format{An object of class \code{character} of length 10.}
\usage{
igraphLib
}
\value{
None
}
\description{
This file sets the required variables for plot
}
\keyword{datasets}
