% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactiveSet.R
\name{interactiveSet}
\alias{interactiveSet}
\title{interactiveSet}
\usage{
interactiveSet()
}
\arguments{
\item{interactiveSet}{A function}
}
\value{
None
}
\description{
This function sets the required D3 variables for the interactive network view.
}
