% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myView.R
\name{myView}
\alias{myView}
\title{Define a view function that will recognize it is in RStudio}
\usage{
myView(x, title)
}
\description{
See the discussion here: https://stackoverflow.com/questions/48234850/how-to-use-r-studio-view-function-programatically-in-a-package.
Turns out that we can't just call View directly, because it needs more
information about the environment it is in.  I won't export this, cuz it will just get used
within a function in this package.
}
