% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{trial}
\alias{trial}
\title{Trial Function}
\usage{
trial()
}
\arguments{
\item{trial}{is a function}

\item{readline}{reads the user's input}

\item{prompt}{prompts the user with a question/warning etc.}
}
\value{
the graph
}
\description{
This function lets user decide what kind of graph \cr
they want to use, for now it is limited to \cr
plot, interactive networks and sankey network
}
\examples{
trial()

}
