% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useShiny.R
\name{useShiny}
\alias{useShiny}
\title{Shiny}
\usage{
useShiny()
}
\value{
Shiny Representaion
}
\description{
This function lets user work with Shiny.
It can be called from draw.R.
User can use the function draw() and type 'shiny' \cr
to answer the question to activate Shiny.
Shiny is set to show only interactive and Sankey networks.
}
