% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{data_load}
\alias{data_load}
\title{Load data}
\usage{
data_load()
}
\value{
None
}
\description{
This function asks user to specify the location of the data for edges and
for nodes separately in their local computer and reads the specified CSV
formatted data into R if user doesn't specify a location for the data,
example data will be loaded.
Please answer the questions with their full location like:
"/Users/JaneDoe/Desktop/BCB410/Rview/inst/extdata/data_edges.csv"
}
