% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataLoad.R
\name{data_load}
\alias{data_load}
\title{Load data}
\usage{
data_load()
}
\value{
None
}
\description{
This function loads the example data.
If user replaces example data with the desired data, it will load user's data.
}
