% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{trial}
\alias{trial}
\title{Trial Function}
\usage{
trial()
}
\arguments{
\item{trial}{Is a function}
}
\description{
This function lets user decide what to draw (plot, interactive networks, sankey network)
}
\examples{

}
\keyword{interactive}
\keyword{network}
\keyword{plot}
\keyword{sankey}
